+++ openwrt-feed/luci-app-fwlive/htdocs/luci-static/resources/fwlive/logging.js
@@ -262,6 +262,8 @@ function buildEmptyStateNodes(state, callbacks) {
 		);
 	}
 
+	if (!st) return nodes;
+
 	if (blocker === 'no_wan_zone') {
 		nodes.push(
 			E(